requirements.txt
Flask
Flask-Cors
google-generativeai
python-dotenv
numpy==1.26.3
scikit-learn==1.3.2
streamlit==1.29.0
streamlit-option-menu==0.3.6